Add schema directory
